src-backend/pyproject.toml
[project]
name = "src-backend"
version = "0.1.0"
description = "RAG Backend with LlamaIndex and Qdrant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "openai>=1.50.0",
    "llama-index-core>=0.11.0",
    "llama-index-vector-stores-qdrant>=0.3.0",
    "llama-index-llms-openai-like>=0.2.0",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-readers-file>=0.2.0",
    "qdrant-client>=1.11.0",
    "sse-starlette>=2.1.0",
    "pydantic-settings>=2.5.0",
    "najia>=2.0.1",
    "arrow>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
